Stop defining SIOC{S,G}ETVLAN as SIOC{S,G}IFGENERIC because they are two differents ioctls manipulating structures of a different size.
In particular this prevents vlan(4) to handle pppoe(4)-specific
ioctls and vice versa, that could lead to smashing the stack of
the caller.

Issue originally found by David Hill when compiling ifconfig
with clang.

Note that this is an ABI break but we're riding a biger dragon
right now.

ok mikeb@, kettenis@, sthen@