Implement POSIX functionality; revise STANDARDS in ps.1.
Add group selection flag ([-G groups]). Convert -ptU to expect a comma
or space separated list instead of a single argument. Let the -U flag
accept user IDs in addition to user names and let it select processes by
real ID, as required by POSIX, instead of the effective user ID.

Make clear that ps doesn't implement the XSI extension. The behaviour of
the -e flag doesn't collide with POSIX.1 but with XSI, so do -dlu. The
flags -edvWwx are extensions to POSIX.1.